Add files via upload
work-in-progress map file with quite a bit of de-activated code (in green). this could be erased for now for sure. this file also unfortunately contains all the data for the Toronto wards' layer. This data is here because, if i understand correctly, i would have had to bypass the CORS protocol or understand what it needed to have the data file kept separately.