Add files via upload
all these files are necessary for precarious_layers_PLAY_August82022.html to work, except some of the images. the one not useful are images i used in past versions of the map. i could sort them out and delete them. the important thing to know is that in a version that works, landacknowledgement.html + ourprivilege.html + whyamap.html are in a folder called LEFTNAV, and, all images are in a folder called markers.